server/requirements.txt
django>=1.11,<1.12
django-pipeline>=1.6,<1.7
ims-bootstrap
django-markdownx